+++ src/bin/email_api_dir/monitoring_handlers/mod.rs
@@ -1,0 +1,11 @@
+// monitoring_handlers — split into submodules in cycle 26 (LOC reduction)
+// Public API preserved via `pub use` re-exports; parent uses `pub use monitoring_handlers::*`.
+
+mod monitoring;
+mod security;
+mod shared;
+
+pub(crate) use monitoring::*;
+pub(crate) use security::*;
+#[allow(unused_imports)]
+pub(crate) use shared::*;
